Simplify module loading.
Rather than tracking names to modules and then later looking them up in a library, put all the relevant bits of info inside of the new ModuleDef class, and use methods on it to load modules. This replaces the old 'library' concept, and significantly simplifies code in other parts of the program. This is a no-op in terms of functionality.

There are a few more TODOs to take care of around simplifying the instantiation procedure, but overall, this is a big improvement.