Rejigger clientdisplaystrategy to be more general.
The original interface was too specific to the staticdisplaystrategy; namely, it
expected a draw to occur as soon as new content was chosen from the load
strategy. This isn't feasible when we might not want to show something
immediately or when we want to animate something. The init/draw pair is more
generally useful.