Remove register from the sandbox.
Replace it with a real global that figures out which module its dealing with via
a side-channel or by looking at the current script.

Turns out, abusing require for not-really-global deps doesn't work quite the
way I wanted. Rather than doing that, I'll be swapping to a locator pattern,
which will be passed to the module_interface in the constructor.

However, that locator pattern doesn't work for register, because we don't even
call the constructors until after we want to have registered. So, we have to
make register work in both the current fake-require environment and the eventual
real requirejs environment. Whew!