Revert "Switch per-module requires to service locator."
This reverts commit 609d71c74a8d09228e3077661ef1682d9237a636.

Instead, use the fact that we are rewriting the modules we are serving anyway to
include a 'define' wrapper. THIS module is created in the right requirejs
context, obviating the prior issues with the requirejs global. This also allows
the module writer to use require to use globals correctly and avoid the wordy
service locator pattern.