Fix over-eager parsing of deps in requirejs.
Sometimes, we want to conditionally load modules. In the server context, this
is no big deal, because we just guard the require by some control flow. On the
client, this doesn't work, because the client has to load its deps
asynchronously. requirejs works around this problem by using a regex to find
all uses of 'require(somepath)' in source, loading those, and then executing the
source. This means that a server-only dep like server/util/googleapis will be
discovered by this regex, and requirejs will attempt to load it.

To fix this, introduce a serverRequire function that is in every way identical
to require, except that it's not found by the regex.