Introduce a new state machine class.
It's simpler than the old one, and fixes a terrible bug where 'external'
transitions, such as ones that request a layout change, might be lost if
sent on the same tick as some internal transition.

This new class is temporarily introduced next to the existing state machine, so
I can slowly convert over the existing machines. When that's finished, I'll
delete the original and move the references.