Fix memory leak in server/fake_require
The "easy" fake require turned out to have a weird leak. The line:

    let ret = path => require(path);

created a for-realz require call that server modules can use to load other
server code. But this require is bound to fake_require's module itself, as it's
the same require that fake_require uses to load its dependencies. Because
modules are loaded in a tree, the server module's deps were being added to the
child list of fake_require. Oops! Simple fix: prune that list, let gc handle the
rest. Tested by manually examining profiles.