Replace handleError API with errorlistener in sm.
Our eventual goal is to have some external system driving the state machines, rather than the state machines driving themselves to the next module. Part of this is recovering correctly when an error occurs. The current implementation of error handling forces each sm to know about how to recover. The new error handling allows other systems (such as parent sms) to be told when an error occurs and to tell that state machine to do something about it.

This actually fixes a potential bug that I hadn't considered before now where a single client sm might have gotten screwed up, but the wall wouldn't have advanced the whole wall to the next module. Now, we theoretically handle this correctly. I'm not sure if this even happened in practice, though.