Drive SMs according to the playlist.
This moves all of the logic for moving from one module to another, or one layout to another outside of the state machines themselves. This is pretty major change from the way that we've done all such transitions up to this point, and so this code should be treated very suspiciously. It finally allows us to have more complicated logic around how the playlist works, because rather than needing to encoding the understanding of the playlist format into the SMs, which is an encapsulation violation, we can stick all of that logic in an external 'playlist driver' that knows about how to understand the playlist format. Finally, we have a change of making that driver able to change things based on time of day, some calendar-based schedule, or really, anything else, without having to modify the SMs themselves, which are already too complicated.

Because this changes the API of the two main state machines, layout and module, we also need to update their main callsites: control, where we match the current API that the status page expects; and server, where we invoke the new playlist driver to drive the machines.

One major last piece of work that we need to do to really make our modules smarter is to introduce the notion of a module being able to send an event denoting that it is "finished" for modules that have a certain built-in duration, such as those that play videos. This could be implemented as a promise or simply as an event that the SMs emit and the playlist driver listens for in order to proceed to the next module.

Another janitorial task is to fixup all of the nouns called "deadline" or "time" to be more specific about what these times are referring to. I think if we request a module to play for 10 seconds that we aren't exactly letting the module play for 10 seconds, and this bug is super-hard to track down given our reuse of those two words all over the place.