Move tracking of clients from lsm to server.
This is logical because there should just be a single, global list of clients that the server tracks, not one per lsm as there was before this change. This simplifies lsm quite a bit, which is nice because I want to remove it, and allows me to remove the duplicate client tracking that the msm had to do, significantly simplifying *that* code, too.