Refactor playlist driver to be more class-like.
This removes a TODO there, and makes the driver more class-like. The benefit is
that now it's easy to add other methods, and make the driver the only way to
drive the state machines (rather than special APIs on the lsm). This should make
it easier to eventually remove the lsm entirely.