Move playModule responsibility to driver.
This fixes a bug whereby the user would say to transition to a specific module, but then a timeout would fire and the driver would attempt to transition to the next one in the playlist. By making the driver the sole authority of which module is up next, we avoid this kind of bug.