Replace getcurrentmoduleinfo with driver-based API.
This is *slightly* less useful than what had been there, but it removes the last
dependency on the lsm from control. At this point, the lsm is entirely under the
operation of the driver, which means it should be far simpler to try to remove
the lsm, as all of its deps are in one place.