Upgrade socket.io to latest.
Besides bringing us up to date with head, this also adds some shiny new features. See https://socket.io/blog/socket-io-2-0-0/ for details.

This version also works with ES6-style module imports, which is a great long-term goal for our
work (though it's uncertain how feasible this is for all of the code...).