Remove scaling from config spec.
This was a thing we created early on when we thought it might be clear to specify clients in terms of integers and then scale those to the right places later. It turns out that this doesn't
really work for a wide variety of walls, so we can remove it.