Upgrade p5 to 0.8.0.
Fix two bugs:
1) Since 0.5.0, transformation applied to the context before it's been
actually added to the DOM aren't saved, which makes the transforms in
setup do nothing. Instead, perform those transformations on every frame
in draw (this has the potential to invalidate certain kinds of caches
in some rendering engines, leading to slowdowns, but I've not seen any
of that in testing).
2) We don't actully use WEBGL mode in p5. Rather than rewriting all of
the p5 modules to work under WEBGL mode, remove our attempt use it. In
the future, this should probably be an option that a sketch could choose
via some template-method pattern or whathaveyou.

Tested via p5_test, life, and cosines.