Use module player pattern on server.
This entirely replaces the two remaining state machines on the server,
and, in fact, all of the state machines remaining in the framework. The
new code is far simpler to reason about, and has the benefit of being
trivially driven from outside (in this case, the playlist driver) in a
robust fashion so that soon, we could move the playlist driving to a
separate process.