Add multiplexing library.
This library changes introduces a 'manual' multiplexing system to a
socket.io socket. The wall used to use socket.io namespaces to ensure
that network messages would not spill between modules, but this has some
overhead that we don't want to deal with and some drawbacks that are
outlined in the next commit.

This library is also tested.