Multiplex per-module network on a single connection.
This completely changes the way that per-module networking works.
Before, we relied on socket.io namespaces to ensure that no module could
accidentally communicate with another module that was running at the
same time. This worked, but had three major drawbacks:

1. Relied on internal, undocumented socket.io APIs to ensure proper
   cleanup of namespace sockets.
2. Messages sent before receiver was ready were dropped, which required
   modules to invent complicated mechanisms to ensure both server &
   client were ready before communicating.
3. Initial handshake on namespace caused a noticeable CPU spike on the
   client which dropped some frames.

This change addresses each of those concerns:
1. One socket is opened per server and client, avoiding the need for
   cleanup.
2. Messages are always queued up and processed when the other party
   inits and begins registering handlers.
3. No additional connections are formed, removing the CPU spike.

This changes the API for modules, though, as there is no more
'connection' or 'reconnect' messages (as we maintain a persistent)
connection. Module authors are advised to instead have clients send a
message to the server when ready to receive any critical missives, and
the new infrastructure will ensure proper behavior. To that end, this
commit also updates the existing modules to not depend on these
messages.