Improve performance of slither.
By drawing each snake as a series of bezier paths (and by the
addition of a critical moveto), chrome is able to clip them far more
effectively, even though there are more state changes. This speeds up
the rendering by over 2x.

Also, make the eyes follow an exaggerated direction of the heading,
which makes the snakes feel more alive.