Optimize processPolygon.
In a profile, this was showing up as a very heavy fn. This was because
the old algorithm was asking whether or not each lattice point in the
polygon was inside or not. This is a O(w * h * e) operation, where w and
h are the width and height of the polygon and e is the number of
edges.

Instead, we employ a trick from polygon rasterization to sweep over the
polygon in the +y direction, calculating only 2 intersections per row
for a O(h) algorithm. As you might surmise, this is _significantly_
faster, often achieving a 100x speedup in practice.