Fix remaining tiffany bug and clean up a bit.
Because we might share points between polygons, it's important not to
directly change the value of an existing point (or you could change more
than you intended). Unfortunately, the segmentPolygon routine did just
this. Fixing this assures that newPt remains true for some polygons
send to the client, keeping the segment-drawing intact.

Also, prevent shattering polygons that are in-progress. Unlikely, but
a good idea.

Also also, simplify the cases under which we draw the polys.